feat(precision): switch mixed precision from FP16 to BF16
BF16 has the same exponent range as FP32 (8-bit exponent vs FP16's 5-bit),
eliminating backward-pass gradient overflow that caused ~36% NaN rate with
FP16. The --fp16 flag name is preserved for backward compatibility but now
selects mx.bfloat16 internally.

Changes:
- train_dynamic.py: 6 input casts + compile boundary dtype → mx.bfloat16
- training_waveform.py: GAN waveform view cast → mx.bfloat16
- hardware.py, run_config.py, training_cli_main.py: updated comments/help text
- All 977 tests pass

+++ DeepFilterNet/df_mlx/training_waveform.py
@@ -100,11 +100,11 @@ def _gan_waveform_view(wav: mx.array, *, use_fp16: bool) -> mx.array:
     """Return GAN discriminator waveform view in the desired precision.
 
     GAN discriminator activations are a major memory contributor when adversarial
-    training activates. Keeping this path in model precision (FP16 when enabled)
+    training activates. Keeping this path in model precision (BF16 when enabled)
     reduces peak memory while MRSTFT can still run in FP32 for stability.
     """
-    if use_fp16 and wav.dtype != mx.float16:
-        return wav.astype(mx.float16)
+    if use_fp16 and wav.dtype != mx.bfloat16:
+        return wav.astype(mx.bfloat16)
     return wav
